prepare for linux and windows separation
