+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["rapl-windows"]
+members = ["rapl-windows", "rapl-rust-lib"]
