+++ rapl-rust-lib/src/ffi.rs
@@ -1,8 +1,8 @@
-use crate::rapl::start_rapl_impl;
+use crate::rapl::windows::start_rapl_windows;
 
 #[no_mangle]
 pub extern "C" fn start_rapl() -> usize {
-    start_rapl_impl()
+    start_rapl_windows()
 }
 
 #[no_mangle]
