+++ rapl-rust-lib/src/rapl/windows.rs
@@ -1,0 +1,3 @@
+pub fn start_rapl_windows() -> usize {
+    123
+}
